lp1373203: Allow users to escape metarecord holds
Provides a means for users to escape a metarecord hold if they accidentally
clicked the Advanced Hold Options link. A Basic Hold Options link will be
provided to allow the user to return to a title hold. If the hold starts as a
metarecord hold, the basic hold options link won't display since we don't know
which bib the patron wants.

Also fixes an issue where the hold cancel button always returned the user to
the previous screen. We will instead utilize the existing hold_source_page
parameter to identify the return page.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>