LP#1423813: adjust some queries to account for deleted copy locations
This patch adjusts some queries so that if a copy location
is deleted, any undeleted items that are attached to it
are considered deleted for the purpose of generating
OPAC item lists and seeing if items are potentially available
to fill hold request.

It should be noted that if a copy location is marked as
logically deleted, there shouldn't be any undeleted
copies associated with it.  However, as this condition is not
presently enforced by a database contraint, adding additional
checks on acpl.deleted provides some additional defense.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>