LP#1403967: show 'subject heading thesaurus' value release notes
The 'subject heading thesaurus' basically defines the issuing authority
that created the authority record. For example, a value of "a" means that
it came from LoC. A value of "n" means that it is a locally created
authority. Note that within the authority MARC editor the 'Subject
Heading Thesaurus' is stored/displayed in the 'Subj' fixed field.

Signed-off-by: Yamil Suarez <yamil@yamil.com>