LP#1234220 Improve hold/copy ratio renewal messages
Display friendlier messages in the TPAC and staff client client when a
hold/copy ratio rule prevents a renewal.

This creates top-level ILS events for total/avail hold copy ratios,
since that's what best works with the staff client.  The TPAC displays a
simple "item is needed for a hold" message, whereas the staff client
displays the raw event description, which is slightly more descriptive.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>