LP#1442254: adjust how sending library set for messages created from notes
When creating a new user message from a public note, the library
of the workstation of the staff member who created the original
note is used if available from get_audit_info().  If it's not
available (e.g., if notes are added using direct SQL, the
home library of the note creator is used instead).

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>