LP#1442701: prune 'message_id' and 'single' CGI params from TPAC dashboard links
This patch fixes a bug where if one goes to the single patron
message view in TPAC, clicking on the 'Message' button in the
dashboard button bar does not return one to the list of messages, but
instead takes the user back to the single message the user was
viewing.  This patch also prunes the CGI parameter from other
links in the patron dashboard.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>