LP 1444130: Add max_chunk_size guards to Holds.pm.
Web staff client fixes introduced a bug in using the Clear
Holds Shelf checkin modifier.  Commit 4ccbf980 added calls
to max_chunk_size on the client to use chunked responses for
better feedback during long operations.  Thise works just
fine when a regular OpenSRF AppRequest is used, but leads to
a network error when AppSubrequest is used, as apparently the
xul client does with the Clear Holds Shelf checkin modifier.

This commit adds code to check if the client object can do
the max_chunk_size call when the max_chunk_size call is
made.  This will prevent the network errors in cases where
the client is an AppSubrequest object and will still use
chunked responses when the client is capable of doing it.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>