LP#1391282: Add Auto-Submitted: auto-generated header to included A/T email templates.
This patch updates various stock A/T definitions for
email notifications to add the following email header:

Auto-Submitted: auto-generated

This specifies to email clients that the message was
generated by a computer program and that no automatic
response is desired (particularly out-of-office/vacation
responses).

No upgrade script is provided; administrators who
wish to use this header for existing notices should
update them manually.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>