LP#741788: install the MARC stream importer better
This patch ensures that marc_stream_importer.pl gets
installed in the Evergreen binary directory and that an
example configuration file gets installed in the appropriate
configuration directory (rather than, as in the past,
the Evergreen bindir).

This patch also causes `oils_header.pl`, which is used by
other command-line scripts, to be installed in the bin
directory
by default.

To test
-------
[1] Run a normal Evergreen installation. Verify that
    marc_stream_importer.pl is installed in /openils/bin
    (or the equivalent if you've chosen a non-default installation
    directory) and that marc_stream_importer.conf is
    installed in /openils/conf (or its equivalent)
[2] (optional) If you have access to OCLC Connexion, verify that you
    can run the MARC stream importer and use it to push records into
    Evergreen.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Jason Stephenson <jason@sigio.com>