LP#1452950 angularize patron registration phase I
Replace legacy Dojo patron registration / edit UI's in the browser
client with an initial cut of an Angular version.  For this commit, the
UI is basically a wireframe, but the selectors display values and
most fields display the correct values set on the patron.

No save or clone etc. operations or data validation are functional.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>