LP#1452366: allow EGWeb context loaders to have child_init actions
This patch adds the ability for EGWeb to call
an initialization function for context loaders
during the child_init phase of Apache backend
startup.

In particular, portions of the R/O object cache
for EGCatLoader are now initialized when a backend
starts up; testing indicates that this can shave
a couple seconds off the time it takes for a
backend to render a bib details page the first time.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>