LP 1198465: Set restored overdue timestamp to time of last overdue
When we have our settings configured to generate new overdues on lost
item return, we start generation after the most recent overdue fine.
Because of this, we need the restored fine to be dated in the past,
which in turn allows the fine generator to apply catch-up fines as
expected.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>