LP#1456021 Enable apache modcgi during install for Debian Jessie
Apache mod cgi isn't enabled when installing apache2.4 on Debian Jessie,
this causes the /updates/manualupdate.html to not be processed and
to just return the raw code of that script since it uses mod cgi.

This adds 'cgi' to the list of apache modules to enable during the
dependency install.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Ben Shum <bshum@biblio.org>