LP#1464709 Seamless non-standard copy status checkout
Add a new boolean configuration option to copy statuses which allow such
copies to be checked out without requiring an alert or override.

New column is config.copy_status.checkout_ok.  Any status with this
boolean enabled will be treated the same as Available, Reshelving, and
Holds Shelf status copies in that the status will be considered
checkout-able.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>