LP#1289486: Remove blocking constraint from audit table
The ACQ lineitem audit table was augmented with an fkey pointing
to the queued record that was related to the line item.  That's
bad because there are also triggers which attempt to manage that
table.  Hilarity ensues.  We should not have an fkey on an audit
table, so, we now remove that.

In order to test this, first attempt to delete an ACQ queue and
have that fail.  Apply the DB change, attempt again, and have it
succeed.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>