LP#1154656 MARC Expert Search "Add Rows" adds duplicate row
This patch changes the behavior of the "Add Search Row" link
(JavaScript function from simple.js "addExpertRow()") so that
we are always cloning the initial "empty" row instead of
potentially cloning a row -- by way of cloneNode() -- with user
input in place, and causing duplicate search rows.

To reproduce the bug, go to Advanced Search, fill in all three
search fields with some text, click on Add Search Row.  You wil
see that the Title search row along with the data in the text
box is duplicated and added to the search.

After this patch, the new row should be added with an empty
text box.

Signed-off-by: Michael Peters <mpeters@emeralddata.net>
Signed-off-by: Rashma Kumaran <rkumaran@kcls.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>