LP#1396263: Fix for 'Return to Record' link breaks after texting
If user is already logged into OPAC when texting a call number from
the record screen, the 'Return to Record' link works properly. However,
if user is not logged in and is prompted to log in first, the 'Return to
Record' link actually returns the user to 'My Account.' This patch
corrects that problem and returns the user to the record regardless of
whether they were initially logged in or not.

Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>