LP#1467634: Hide saved searches from screen readers for non-staff client
This commit will only allow screen readers to see the saved searches for
users that are logged through XUL client.
Assuming that a screen reader can be attached to the XUL client.

It should also work for a user logged in through the web staff
client that is using a screen reader.

To test
-------
[1] Apply the patch, then perform a search in the public catalog and
    view the HTML source.  Verify that there's no "sr-only" heading
    for "Saved searches".
[2] Set the opac.staff_saved_search.size library setting to a non-zero
    value.
[3] In the web staff client, perform a search using the embeded OPAC;
    verify that the recent searches list is displayed.
[4] Repeat step 4 in the XUL client.

Signed-off-by: Yamil Suarez <yamil@yamil.com>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>