LP1464765: Speed Up Number Padding for Sort Keys
[New commit message by Galen Charlton:]
This patch was originally written to fix a problem with
normalization of various labels, but the patches for
bug 1155313 fixed the function problem.  However, testing
shows that this version of evergreen.lpad_number_substrings()
is about a third faster, so this patch is being merged to
get the speed improvement and the additional tests.

Thanks to Thomas Berezansky for a much simplified approach.

Signed-off-by: Jason Boyer <jboyer@library.in.gov>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>