LP#1470957: Normalize date1 (pubdate sorter) to null when not numeric
We supply a normalizer function for making sure date-ish values are
always numeric (several, in fact) but we don't enable them by default
for appropriate fields.  This does that.

Additionally, we allow NULL sorters so that the reingest of a "broken"
record will get fixed.

We avoid the need for a reingest to cover the sorter issue by removing
non-numeric pubdate sorters.

To test
-------
[1] Identify some records where the Date1 value in the 008 field
    is '||||'.
[2] Perform a catalog search that includes some of those records
    and sort by publication date newest to oldest. Note that
    the |||| records sort at the top.
[3] Apply the patch and perform the schema update.
[4] Perform the search again (possibly clearing the search
    results cache first).  This time, the |||| records should
    sort near the end.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Jennifer Pringle <jennifer.pringle@bc.libraries.coop>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>