webstaff: Fixed-position navigation menu for web client
Changes the class for top-navbar so that it stays fixed at the top of the
screen. This is useful for staff who will need to quickly jump to another
area of the client after they have scrolled down a page.

Initially, changing the class shifted some of the content higher so that the
navbar was covering some elements. I addressed this by adding some padding to

Many thanks to Jason Etheridge for pointing me in the right direction to get
started.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>