LP 1198465: Load negative balance test transactions in load_all.sql
This commit integrates the SQL needed to setup negative balance
testing into the load_all.sql which loads the rest of the test data.

It also separates out the "reset" commands into a separate SQL file.
They are useful when doing repeated testing, but unnecessary and
may eventually require manual tweaking, so they are provided here as a
convenience only.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Remington Steed <rjs7@calvin.edu>