LP#1479953: Rename indexes to *_idx and add pgTAP test.
To make the new index names more consistent with existing indices,
I took the liberty of adding _idx to the new index names.

We also add a pgTAP test to verify that the new indices exist in
the database.  The test would fail before running the upgrade
script and succeed after.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>