LP 1484655: ftp://ftp.mozilla.org moved to http://archive.mozilla.org
This commit changes the XULRUNNER_URL in Open-ILS/xul/staff_client/
Makefile.am from using ftp to talk to ftp.mozilla.org to using http
with archive.mozilla.org.  This is necessary because Mozilla.org
have apparently declared the former to be closed and the latter to
be the canonical location for downloads.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>