webstaff: add support for editing authority records
This adds basic support for invoking the MARC editor on
authority records, and fixes a couple bugs discovered
along the way. The route currently supported is

/cat/catalog/authority/:authority_id/marc_edit

In the future, some sort of summary view of an authority
record might be added, in which case the route
"/cat/catalog/authority/:authority_id" is available.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>