LP#1489955: enable filtering authority search/browse by thesaurus
This patch adds support for restricting the scope of an
authority headings browse or authority record search by
thesaurus.  For example, in SuperCat, the URL pattern

/opac/extras/browse/marcxml/authority.subject/1/heart/0/20

generates returns a browse of up to 20 authority records
centered around the subject heading "heart".  With this patch,
the URL pattern

/opac/extras/browse/marcxml/authority.subject/1/heart/0/20/c

restricts the browse to MeSH headings. Multiple thesauruses
can be specified as well; for example, this URL pattern

/opac/extras/browse/marcxml/authority.subject/1/heart/0/20/a,b

browses in LCSH and LC Children's.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>