LP1490616: Adjust "Penalties & Messages" label (webby)
The pane in the web-based staff client to display penalties and messages
currently has a label of "Staff-Generated Penalties/Messages", which is
misleading as the list includes system-generated penalties and messages.

Modify this to "Penalties and Messages" to avoid the "/" construction
(which is considered harder to translate) and avoid the impression that
it contains only staff-generated penalties and messages.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <ben@evergreener.net>