lp1494427: Fix refund error
Fixes an error that occurs when attempting to refund a negative balance from
the patron's bill record. We removed the staff.patron.bills.handle_refund.
confirm_message string, which caused the error when the handle_refund function
tried to retrieve the string. It turns out the string never displayed in the
client. Therefore, in this branch, we stop the handle_refund function from
trying to retrieve the string. I did the confirmation message for the
adjust_to_zero function since the same issue was occurring there.

Test case: Find a negative balance transaction and, from the patron's bills
record, select the transaction and select the option to Refund. Pre-patch, this
action will generate an error message. Post-patch, this action should generate
a prompt asking the user to confirm the action.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>