LP#1340852: Retain search params for copy location groups
The order in which we determined what to 'selected' in the "Library"
dropdown put the resolved locg first (which allows for shortnames, e.g.
BR1).  That only gives us half the location group though, which is why
it was bumping up a level.  Now, if we have a location group selected,
it will use that first (since they don't have shortnames); the remaining
select order will be the same.

Signed-off-by: Jake Litrell <jake@masslnc.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>