Docs 2.9: Update and add detail for Holdings Import profile
In order to add details about the new ability to import Stat Cat data
for a copy, this commit adds a table of all the import fields available
in the Holdings Import Profile. It also duplicates some content from
another section about batch importing, and updates the screenshot and
some wording in the original section.

Signed-off-by: Remington Steed <rjs7@calvin.edu>