LP#1463166 Fix Batch update of Funds in Selection Lists
Selection lists are currently unable to batch update funds.  The problem
stems from the fact that the Fund selection list is not populated in the
copy rows.  New code used to limit the number of funds, in the dropdown
list, to those funds available to the user was added for purchase
orders.  However, these changes were not added to Selection Lists.  This
code adds the new call to initBatchUpdater to the selection list code.

Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>
Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>