webstaff: give more feedback for Z39.50 edit-then-import
The "edit then import" modal now varies the label of the
MARC editor's save button based on whether one has yet
to import the current record or if one is just making
changes to it after having imported it. The dialog now
also has a button for going to the newly-imported record
in catalog view.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>