LP 1499123: Add ignore_proximity to config.standing_penalty.
This commit adds the integer column ignore_proximity to the
config.standing_penalty table.  It also adds the column to the
csp class entry in the IDL.

It also modifies the action.hold_permit_test() function from
110.hold_matrix.sql to use the ignore_proximity field from
config.standing_penalty when checking the user's penalties to see if
they block the hold.

We also modify the action.item_user_circ_test() function from
100.circ_matrix.sql to use the ignore_proximity field from the
config.standing_penalty table when checking to see if the user's
penalties block the circulation.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>