LP#1501781 - Make patron name search diacritic/space insensitive.
Diacritical marks may exist in the patron record, as they should, but
this makes patron name search difficult for librarians who may be
unfamiliar with all the special characters used and also how to elicit
them from a keyboard.  To ease this, accented characters will be
converted into their 'plain' analogs for comparison purposes.

So, for example, if the patron's last name is Chávez, typing "Chavez"
in the Last Name box in Patron Search will match it.

Spaces in a name (like "De la Croix") will be squashed out so that
name would be matched by "Delacroix" or "De la Croix" or variants.

The librarian can enter the accented characters or not.

Signed-off-by: Dan Pearl <dpearl@cwmars.org>
Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>