LP 1502156: Fix marc_export error when dumping authorities.
There is a chance that marc_export will attempt to call a nonexistent
field on MARC::Record if an error occurs while exporting authority
records.  The bug could only be triggered if a conversion error
occurred while exporting the authority as either MARCXML or USMARC.

The fix is to rename a couple of variables in the Marque::Authority->next
method to be more like those used in Marque::Biblio->next.  This will have
the side effect of making marc_export easier to maintain, since the
variables used in one method will now have the same meaning as those
in the other.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>