LP#1197636 - Email record detail does not check for email
Even though no email was configured, the message shown when requesting
bib information to be emailed indicated success.  This misleading
behavior has been addressed by alerting the patron of the situation
and providing a link to the page where the email address is defined,
after which the action is automatically retried.

The success message has also been altered slightly to provide
confirmation of the generated email.

To test:

[1] Create a patron account that does not have an email address.
[2] In the OPAC, log in as that patron and attempt to email a
    bib record.  A message should display indicating that the
    email can be sent; that message should contain a link to the
    user's account page where they can set their email address.
[3] After setting the email address, you should then be
    redirected back and the email sent.

Signed-off-by: Dan Pearl <dpearl@cwmars.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <ben@evergreener.net>