LP#1503782: fix horizontal scrollbar in MARC import queue inspector
In the XUL client (and webstaff), attempting to view the contents
of a MARC record import queue displays a table that, due to
its width, will almost always have a horizontal scrollbar. However,
attempting to actually use the scrollbar often causes the
focus to jump to the top of the table instead.  This patch
fixes that.

To test:

[1] Create a bib queue that has at least a dozen records or so.
[2] Inspect the contents of the queue, scroll to the bottom of
    the table, and interact with the horizontal scrollbar.
[3] Observe that doing so typically results in jumping up
    to the top of the table.
[4] Apply the patch, clear cache and refresh, and try step #2
    again.
[5] This time, the horizontal scrollbar should act normally.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>