webstaff: start teaching egHatch about sessionStorage
This patch adds getSessionItem(), setSessionItem(), and
removeSessionItem(), which are all wrappers around
$window.sessionStorage. This is done to support settings
whose values are sticky for the duration of a session,
which at present is the lifetime of the browser tab.

An alternative would be using session cookies; some user
testing is needed to see which lifetime makes the most
sense.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>