LP 1503867: Use ws_ou in open-ils.circ.copy_transit.receive.
After putting NCIPServer into production, we discovered that a staff
user cannot receive a transit destined for their work org. unit if
the actor.usr.home_ou is not the same as their current work_ou.
This occurred because transit_receive in OpenILS::Application::
Circ::Transit uses the requestor's home_ou and not the ws_ou.

If you want to trigger this bug, write a little script to receive
transits, then login with a staff account at a location other than
their home_ou and try to receive a transit destined for the work_ou.
It will fail.

This small commit replaces home_ou in two places with ws_ou in order
to make the above work.

NOTE: This bug affects an API call not used in Evergreen except in
one test script.  It used in two ILL add-ons for Evergreen: issa and
NCIPServer.  Receiving transits by checking copies in via the staff
client, the typical workflow, is not affected by this bug or its fix.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>