LP#1507845: Improve search engine optimization
Back in bug # 1414033 , we added rel="nofollow" attributes to many of the links
that led to search results or call number browsing. This was the right idea,
but a bit overzealous as it turns out. We actually want search engines to
follow those links; we just don't want search engines to index them.

Therefore, we can tweak the <meta> tag in the results and browse page headers
to say "follow,noindex", and remove many of the inline rel="nofollow"
attributes from the other pages.

In addition, add a rel="canonical" link to the catalogue home page.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <ben@evergreener.net>

Conflicts:
	Open-ILS/src/templates/opac/parts/record/subjects.tt2