LP#1508477: bump up minimum version of angular-hotkeys
Version 1.3.0 of angular-hotkeys is now required, as this
is the version that introduces the ability the specify
that hotkeys can be used even when inputs, selects,
and textareas have focus.

After applying the patch (and doing a bower update,
grunt uglify, and install), one can test hotkeys in
the web staff client by (for example):

- starting from the home page and hitting F1 to
  get to patron search by barcode
- hitting F5 to get the item status page

Note that if the page embeds the OPAC or a Dojo
interface using an egEmbedFrame, if any element
of the iframe's content has focus, the hotkeys won't
work -- that will be the subject of another patch.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>